+++ lib/util/mixins/use_sprite.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:bonfire/base/game_component.dart';
-import 'package:bonfire/util/extensions/extensions.dart';
 import 'package:flame/sprite.dart';
 
 ///
@@ -22,11 +21,11 @@ mixin UseSprite on GameComponent {
   void render(Canvas canvas) {
     super.render(canvas);
     if (this.isVisible) {
-      sprite?.renderWithOpacity(
+      sprite?.render(
         canvas,
-        position,
-        size,
-        opacity: opacity,
+        position: position,
+        size: size,
+        overridePaint: paint,
       );
     }
   }