+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Update `tiledjsonreader`.
 - Fix crash in `Acceleration` mixin.
 - Adds `onStop` params in `Acceleration` mixin. It's called when stop for collision or when speed is equals 0 in `stopWhenSpeedZero` setted true.
+- Use `HasPaint` mixin in `GameComponent`.
 
 # [2.9.4]
 - Adds `revive` method in `Attackable` mixin. Now if adds life to stay above 0 it's is revive.